Changes for mobile
