+++ app/[user]/budget/page.tsx
@@ -137,7 +137,7 @@ export default function BudgetPage() {
             }}
           />
         </div>
-        <div className="rounded-md border bg-white dark:bg-background p-3 shadow-sm">
+        <div className="rounded-md border bg-white dark:bg-background p-3 shadow-sm overflow-x-auto"> 
           {allData.length > 0 ? (
             <DataTable columns={columns} data={allData} />
           ) : (
@@ -181,15 +181,6 @@ export default function BudgetPage() {
                     {insights.overBudgetCategories.join(", ")}
                   </span>{" "}
                   {insights.overBudgetCategories.length === 1 ? "is" : "are"} over budget!
-                </p>
-              )}
-              {insights.suggestedReductionCategories.length > 0 && (
-                <p>
-                  • Consider reducing spending on{" "}
-                  <span className="text-yellow-600 dark:text-yellow-400 font-medium">
-                    {insights.suggestedReductionCategories.join(", ")}
-                  </span>
-                  .
                 </p>
               )}
               {insights.nearingLimitCategories.length === 0 &&
